requirements.txt
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.4.2
torchvision==0.18.0
pytorch-lightning==2.4.0
transformers==4.40.2
scikit-learn
scikit-misc==0.3.1
scib
sympy==1.12
tokenizers==0.19.1
scanpy==1.10.3
cellxgene-census==1.15.0
chex==0.1.86
contextlib2==21.6.0
Deprecated==1.2.14
docrep==0.3.2
einops==0.8.0
etils==1.5.2
filelock==3.14.0
flax==0.8.5
humanize==4.10.0
idna==3.7
iniconfig==2.0.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
json5==0.9.25
leidenalg==0.10.2
lightning==2.4.0
lightning-utilities==0.11.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-venn==1.1.1
mdurl==0.1.2
ml_collections==0.1.1
ml_dtypes==0.5.0
mpmath==1.3.0
msgpack==1.1.0
mudata==0.2.4
multipledispatch==1.0.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
numpyro==0.15.2
opt-einsum==3.3.0
optax==0.2.3
orbax-checkpoint==0.6.3
packaging==24.0
PyYAML==6.0.1
reactome2py==3.0.0
rich==13.8.1
safetensors==0.4.3
triton==2.3.0
urllib3==2.2.1
wrapt==1.16.0
pyarrow==14.0.2
dotmap==1.3.30
datasets==2.19.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
mamba-ssm==2.2.2
